refactor: modularize discovery card architecture
