+++ README.md
@@ -17,6 +17,7 @@ Chrome Music Lab の Song Maker にインスパイアされた、教室向けグ
 - **テンポ制御**: 40–200 BPM のスライダー
 - **設定パネル**: テンポ・音色・音域・ブロック数・黒鍵・使う音・リセットを「⚙ せってい」パネルに集約（リセットは確認ステップ付き）
 - **曲の保存・共有**: Supabase に保存し、みんなの曲一覧ページで共有
+- **はっけんカード**: 音の組み合わせやリズムを再生中に発見し、演出とアルバムで集められる
 - **日本語 / 英語切り替え**: 各ページにロケールトグル（localStorage で永続化）
 - **レスポンシブ対応**: デスクトップ・タブレット・スマホに最適化（タッチ操作対応）
 
@@ -35,6 +36,7 @@ Chrome Music Lab の Song Maker にインスパイアされた、教室向けグ
 | パス | 内容 |
 |------|------|
 | `/` | 作曲エディタ |
+| `/discoveries` | はっけんカードのアルバム |
 | `/songs` | みんなの曲一覧（SNS 風フィード） |
 
 保存した曲は `/songs` に表示され、「あそぶ」ボタンで `/?load=<id>` 経由でエディタに読み込まれます。古い形式で保存された曲は読み込み時に現行モデルへ自動移行されます。
@@ -52,16 +54,20 @@ Chrome Music Lab の Song Maker にインスパイアされた、教室向けグ
 src/
 ├── app/
 │   ├── components/   # Header, SettingsPanel, Grid, NotePanel, SaveModal
+│   │   └── discovery/ # 発見演出・カード・アルバムUI
+│   ├── discoveries/  # はっけんカードのアルバムページ
 │   ├── songs/        # みんなの曲一覧ページ
 │   ├── styles/       # 機能別に分割したグローバル CSS（base/header/settings/grid/…）
 │   ├── layout.tsx
 │   ├── page.tsx      # 作曲エディタ（状態の配線とコンポーネント合成）
 │   └── globals.css   # styles/ を @import するエントリ
 ├── core/             # 型定義・純粋関数による状態操作 + 旧データ移行（DOM/Audio 非依存）
+├── discovery/        # 発見カタログ・タイムライン・ルール・獲得判定（純粋ロジック）
+│   └── rules/        # 和音・メロディ・リズム別の検出ルール
 ├── audio/            # Web Audio エンジン（先読みスケジューリング）
 ├── ui/               # 色・グリッド表示・音名表記（ドレミ）ヘルパー
-├── hooks/            # カスタムフック（曲状態+履歴、タップ/ドラッグ操作、ロケール）
-└── lib/              # Supabase クライアント、i18n 翻訳定義
+├── hooks/            # 状態・操作・発見フィードバックを配線するカスタムフック
+└── lib/              # Supabase クライアント、DBリポジトリ、i18n 翻訳定義
 ```
 
 ### 設計上の約束
