+++ src/discovery/catalog.ts
@@ -1,7 +1,4 @@
-import type { DiscoveryDefinition, DiscoveryId } from './types';
-import { DISCOVERY_IDS } from './types';
-
-export const DISCOVERY_CARDS: readonly DiscoveryDefinition[] = [
+export const DISCOVERY_CARDS = [
   { id: 'interval_third', ruleVersion: 1, sortOrder: 10 },
   { id: 'open_fifth', ruleVersion: 1, sortOrder: 20 },
   { id: 'close_tension', ruleVersion: 1, sortOrder: 30 },
@@ -10,7 +7,12 @@ export const DISCOVERY_CARDS: readonly DiscoveryDefinition[] = [
   { id: 'rest_then_burst', ruleVersion: 1, sortOrder: 60 },
   { id: 'rhythm_loop', ruleVersion: 1, sortOrder: 70 },
   { id: 'sustain_contrast', ruleVersion: 1, sortOrder: 80 },
-];
+] as const;
+
+export type DiscoveryDefinition = (typeof DISCOVERY_CARDS)[number];
+export type DiscoveryId = DiscoveryDefinition['id'];
+
+export const DISCOVERY_IDS: readonly DiscoveryId[] = DISCOVERY_CARDS.map((card) => card.id);
 
 const DISCOVERY_ID_SET: ReadonlySet<string> = new Set(DISCOVERY_IDS);
