+++ src/discovery/types.ts
@@ -1,21 +1,6 @@
-export const DISCOVERY_IDS = [
-  'interval_third',
-  'open_fifth',
-  'close_tension',
-  'stepwise_run',
-  'call_and_response',
-  'rest_then_burst',
-  'rhythm_loop',
-  'sustain_contrast',
-] as const;
+import type { DiscoveryId } from './catalog';
 
-export type DiscoveryId = (typeof DISCOVERY_IDS)[number];
-
-export type DiscoveryDefinition = {
-  id: DiscoveryId;
-  ruleVersion: number;
-  sortOrder: number;
-};
+export type { DiscoveryDefinition, DiscoveryId } from './catalog';
 
 export type DiscoveryMatch = {
   cardId: DiscoveryId;
